Added basic structure to repo
Added a basic first draft of repo structure. There is potential to
change to using a global parent dir that is just git and builder files
then a subdir that is the actual package to keep things clean. I will
test it out before version 1 and see what works best in future.