Added fledgling Erray struct and factory
Begging by creating an example struct for CKKS type encryption called
eCKKS that DOES NOT get exported from the package. Instead to aid in
abstraction we enforce the use of the Erray interface from the
NewCKKSErray factory. This means while we may now write functions with
Erray + CKKS in mind we can simply add a new implementation of Erray
using something else such as BFV or any further advancements in the
future. This is only the beggining as Erray needs to be fleshed out with
expected homomorphic function AS WELL AS being serialisable most likeley
in JSON format for easy tansmission across networks and architectures.